requirements.txt
absl-py==2.0.0
affine==2.4.0
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.36.0
Flask==3.0.0
Flask-Cors==4.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.2
pyparsing==3.1.1
gunicorn==21.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rasterio==1.3.9
six==1.16.0
snuggs==1.4.7
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
